% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Function \code{fars_map_state}}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{This is a single integer value with the state number from the input.}

\item{year}{This value is an input value that gives out either a single year or a range of several years.}
}
\value{
Either a message is given out when there are no accidents or a map of the state and its accidents is shown.
}
\description{
This function takes the state number and a specific year as arguments and prints out a map of accidents in the selected state.
}
\examples{
fars_map_state(1,2015)

}
