% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Function \code{fars_read}}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{It is a string that should contain the filename}
}
\value{
This function returns a data table created by the dplyr package.
The package dplyr and readr are needed in order to make this package work.
}
\description{
This is a function that tries to read in a specified csv file and outputs it as a table
if the file has been found or exits with a warning message.
Within the data object the content of the csv file is written if filename is found.
The option progress = FALSE prevents from displaying a progress bar.
}
\examples{
fars_read(system.file("extdata", "accident_2015.csv.bz2", package = "fars"))

}
