% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function \code{fars_summarize_years}}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{In the function call one ore more years are expected to be submitted.
The dat_list object takes the input of the \code{fars_read_years} function and delivers the years value
to it with the call.
bin_rows comes from the dplyr package and takes the dat_list object and groups it by year and month.
spread comes from the tidyr package and reorganizes the data according to year and the summarized value of n.}
}
\value{
A table with the columns for the months and each year.
}
\description{
This function takes the returned values of other functions and summarizes the values by year and month.
}
\examples{
fars_summarize_years(2014:2015)
fars_summarize_years(2015)

}
