% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Function \code{make_filename}}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Input of a numeric integer value for the year.}
}
\value{
A textstring is returned.
}
\description{
This function takes a year value as input and adds this value into a textstring for the zipped filename.
As this is the last performed operation this is also the return value.
sprintf is a generic C function that pastes a text string (coming from the base package).
The output is created as accident_ followed by the value of the year.
}
\examples{
make_filename(2015)

}
